% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlogistic.R
\name{GenSyntheticLogistic}
\alias{GenSyntheticLogistic}
\title{Generate Logistic Synthetic Data}
\usage{
GenSyntheticLogistic(
  n,
  p,
  k,
  seed,
  rho = 0,
  s = 1,
  sigma = NULL,
  shuffle_B = FALSE
)
}
\arguments{
\item{n}{Number of samples}

\item{p}{Number of features}

\item{k}{Number of non-zeros in true vector of coefficients}

\item{seed}{The seed used for randomly generating the data}

\item{rho}{The threshold for setting values to 0.  if |X(i, j)| > rho => X(i, j) <- 0}

\item{s}{Signal-to-noise parameter. As s -> +Inf, the data generated becomes linearly separable.}

\item{sigma}{Correlation matrix, defaults to I.}
}
\value{
A list containing:
 the data matrix X,
 the response vector y,
 the coefficients B,
}
\description{
Generates a synthetic dataset as follows: 1) Generate a data matrix, 
X, drawn from a multivariate Gaussian distribution with mean = 0, sigma = Sigma
2) Generate a vector B with every ~p/k entry set to 1 and the rest are zeros.
3)  Every coordinate yi of the outcome vector y ∈ {−1, 1}^n is sampled 
independently from a Bernoulli distribution with success probability: 
P(yi = 1|xi) = 1/(1 + exp(−s<xi, B>))
Source https://arxiv.org/pdf/2001.06471.pdf Section 5.1 Data Generation
}
\examples{
data <- GenSyntheticLogistic(n=500,p=1000,k=10,seed=1)
X = data$X
y = data$y
}
