\name{L0Learn-package}
\alias{L0Learn-package}
\alias{L0Learn}
\docType{package}
\title{
\packageTitle{L0Learn}
}
\description{
\packageDescription{L0Learn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{L0Learn}
\packageIndices{L0Learn}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{L0Learn}

Maintainer: \packageMaintainer{L0Learn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
