% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{L0Learn.coef}
\alias{L0Learn.coef}
\title{Extract Solutions}
\usage{
L0Learn.coef(fit, index)
}
\arguments{
\item{fit}{The output of L0Learn.fit}

\item{index}{The index of the solution to extract. A summary of the solutions
and their indices can be viewed by print(fit)}
}
\description{
Extracts a specific solution in the path
}
