% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{L0Learn.cvfit}
\alias{L0Learn.cvfit}
\title{Cross Validation}
\usage{
L0Learn.cvfit(X, y, Loss = "SquaredError", Penalty = "L0",
  Algorithm = "CD", MaxSuppSize = 100, NLambda = 100, NGamma = 10,
  GammaMax = 10, GammaMin = 1e-04, PartialSort = TRUE, MaxIters = 200,
  Tol = 1e-06, ActiveSet = TRUE, ActiveSetNum = 3, MaxSwaps = 100,
  ScaleDownFactor = 0.8, ScreenSize = 1000, AutoLambda = TRUE,
  LambdaGrid = c(0), Nfolds = 10, Seed = 1)
}
\arguments{
\item{X}{The data matrix.}

\item{y}{The response vector.}

\item{Loss}{The loss function to be minimized. The currently supported choice is "SquaredError".}

\item{Penalty}{The type of regularization. This can take either one of the following choices:
"L0", "L0L2", and "L0L1".}

\item{Algorithm}{The type of algorithm used to minimize the objective. Currently "CD" and "CDPSI" are
are supported. "CD" is a variant of cyclic coordinate descent and can run very fast. "CDPSI" performs
local combinatorial search on top of CD and thus can achieve higher quality solutions (at the expense
of increased running time).}

\item{MaxSuppSize}{The maximum support size to reach in the grid before termination. We recommend setting
this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small
portion of non-zeros.}

\item{NLambda}{The number of Lambda values to select (recall that Lambda is the regularization parameter
corresponding to the L0 norm).}

\item{NGamma}{The number of Gamma values to select (recall that Gamma is the regularization parameter
corresponding to L1 or L2, depending on the chosen penalty).}

\item{GammaMax}{The maximum value of Gamma when using the L0L2 penalty. For the L0L1 penalty this is
automatically selected by the toolkit.}

\item{GammaMin}{The minimum value of Gamma when using the L0L2 penalty. For the L0L1 penalty, GammaMin
specifies the fraction of GammaMax at which the grid ends.}

\item{PartialSort}{If TRUE partial sorting will be used for sorting the coordinates (see our paper for
for details). Otherwise, full sorting is used.}

\item{MaxIters}{The maximum number of iterations (full cycles) for CD per grid point.}

\item{Tol}{The tolerance which decides when to terminate CD (based on the relative change in the objective).}

\item{ActiveSet}{If TRUE, performs active set updates.}

\item{ActiveSetNum}{The number of consecutive times a support should appear before declaring support stabilization.}

\item{MaxSwaps}{The maximum number of swaps used by CDPSI for each grid point.}

\item{ScaleDownFactor}{This parameter decides how close the selected Lambda values are. The choice should be between
strictly between 0 and 1 (i.e., 0 and 1 are not allowed). For details, see our paper - Section 5 on Adaptive Selection
of Tuning Parameters).}

\item{ScreenSize}{The number of coordinates to cycle over when performing correlation screening.}

\item{AutoLambda}{If FALSE, the user specifier a grid of Lambda0 values through the Lambda0Grid parameter. Otherwise,
if TRUE, the values of Lambda0 are automatically selected based on the data.}

\item{LambdaGrid}{A vector of Lambda0 values to use in computing the regularization path. This is ignored unless AutoLambda0 = FALSE.}

\item{Nfolds}{The number of folds for cross-validation}

\item{Seed}{The seed used in randomly shuffling the data for cross-validation}
}
\value{
An S3 object of type "L0Learn" describing the regularization path. The object has the following members:
\item{a0} {For L0, this is a sequence of intercepts. Note for L0L1 and L0L2, a0 is a list of intercept sequences,
where each member of the list corresponds to a single gamma value.}
\item{beta} {For L0, this is a matrix of coefficients of dimensions p x \code{length(lambda)}, where each column
corresponds to a single lambda value. For L0L1 and L0L2, this is a list of coefficient matrices, where each matrix
corresponds to a single gamma value. }
\item{lambda} {For L0, lambda is a sequence of lambda values. For L0L1 and L0L1, it is a list of lambda sequences,
each corresponding to a single gamma value.}
\item{gamma} {For L0L1 and L0L2, this is a sequence of gamma values.}
\item{suppsize} {For L0, this is a sequence of support sizes (number of non-zero coefficients). For L0L1 and L02,
it is a list of support size sequences, each representing a single gamma value.}
\item{converged} {For L0, this is a sequence indicating whether the algorithm converged at the current point in the
regularization path. For L0L1 and L0L2, this is a list of sequences, each representing a single gamma value.}
}
\description{
Fits an L0 model on the full data and performs K-fold cross-validation.
}
