% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{L0Learn.predict}
\alias{L0Learn.predict}
\title{Prediction}
\usage{
L0Learn.predict(fit, x, index)
}
\arguments{
\item{fit}{The output of L0Learn.fit}

\item{x}{The sample which can be a vector or a matrix}

\item{index}{The index of the solution to use for prediction. A summary of the solutions
and their indices can be viewed by print(fit)}
}
\description{
Predicts the response for a given sample
}
