% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{print.L0Learncoef}
\alias{print.L0Learncoef}
\title{Print L0Learn.coef object}
\usage{
\method{print}{L0Learncoef}(x, ...)
}
\arguments{
\item{x}{L0Learn.coef object}

\item{...}{ignore}
}
\description{
Prints a summary of L0Learn.coef
}
