% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.L0Learn}
\alias{coef.L0Learn}
\title{Extract Solutions}
\usage{
\method{coef}{L0Learn}(object, lambda, gamma = 0, ...)
}
\arguments{
\item{object}{The output of L0Learn.fit}

\item{lambda}{The value(s) of lambda at which to extract the solution.}

\item{gamma}{The value of gamma at which to extract the solution. Note that, unlike lambda, this can only take single values.}

\item{...}{ignore}
}
\description{
Extracts a specific solution in the regularization path
}
