% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{L0Learn.fit}
\alias{L0Learn.fit}
\title{Fit an L0-regularized model}
\usage{
L0Learn.fit(X, y, Loss = "SquaredError", Penalty = "L0", Algorithm = "CD",
  MaxSuppSize = 100, NLambda = 100, NGamma = 10, GammaMax = 10,
  GammaMin = 1e-04, PartialSort = TRUE, MaxIters = 200, Tol = 1e-06,
  ActiveSet = TRUE, ActiveSetNum = 3, MaxSwaps = 100,
  ScaleDownFactor = 0.8, ScreenSize = 1000, AutoLambda = TRUE,
  LambdaGrid = c(0))
}
\arguments{
\item{X}{The data matrix.}

\item{y}{The response vector.}

\item{Loss}{The loss function to be minimized. The currently supported choice is"SquaredError".}

\item{Penalty}{The type of regularization. This can take either one of the following choices:
"L0", "L0L2", and "L0L1".}

\item{Algorithm}{The type of algorithm used to minimize the objective. Currently "CD" and "CDPSI" are
are supported. "CD" is a variant of cyclic coordinate descent and can run very fast. "CDPSI" performs
local combinatorial search on top of CD and thus can achieve higher quality solutions (at the expense
of increased running time).}

\item{MaxSuppSize}{The maximum support size to reach in the grid before termination. We recommend setting
this to a small fraction of min(n,p) (e.g. 0.05 * min(n,p)) as L0 regularization typically selects a small
portion of nonzeros.}

\item{NLambda}{The number of Lambda values to select (recall that Lambda is the regularization parameter
corresponding to the L0 norm).}

\item{NGamma}{The number of Gamma values to select (recall that Gamma is the regularization parameter
corresponding to L1 or L2, depending on the chosen penalty).}

\item{GammaMax}{The maximum value of Gamma when using the L0L2 penalty. For the L0L1 penalty this is
automatically selected by the toolkit.}

\item{GammaMin}{The minimum value of Gamma when using the L0L2 penalty. For the L0L1 penalty, GammaMin
specifies the fraction of GammaMax at which the grid ends.}

\item{PartialSort}{If TRUE partial sorting will be used for sorting the coordinates (see our paper for
for details). Otherwise, full sorting is used.}

\item{MaxIters}{The maximum number of iterations (full cycles) for CD per grid point.}

\item{Tol}{The tolerance which decides when to terminate CD (based on the relative change in the objective).}

\item{ActiveSet}{If TRUE, performs active set updates.}

\item{ActiveSetNum}{The number of consecutive times a support should appear before declaring support stabilization.}

\item{MaxSwaps}{The maximum number of swaps used by CDPSI for each grid point.}

\item{ScaleDownFactor}{This parameter decides how close the selected Lambda values are. The choice should be between
strictly between 0 and 1 (i.e., 0 and 1 are not allowed). For details, see our paper - Section 5 on Adaptive Selection
of Tuning Parameters).}

\item{ScreenSize}{The number of coordinates to cycle over when performing correlation screening.}

\item{AutoLambda}{If FALSE, the user specifier a grid of Lambda0 values through the Lambda0Grid parameter. Otherwise,
if TRUE, the values of Lambda0 are automatically selected based on the data.}

\item{LambdaGrid}{A vector of Lambda0 values to use in computing the regularization path. This is ignored unless AutoLambda0 = FALSE.}
}
\value{
An object of type "L0Learn" containing all the solutions in the computed regularization path.
}
\description{
Computes the regularization path for the specified loss function and
choice of regularization (which can be a combination of the L0, L1, and L2 (squared) norms).
}
