% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{plot.L0Learn}
\alias{plot.L0Learn}
\title{Plot Cross-validation Errors}
\usage{
\method{plot}{L0Learn}(cvfit, ...)
}
\arguments{
\item{cvfit}{L0Learn.cvfit object}

\item{...}{ignore}
}
\description{
Plots cross-validation errors
}
