% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{predict.L0Learn}
\alias{predict.L0Learn}
\title{Predict Response}
\usage{
\method{predict}{L0Learn}(fit, newx, lambda, gamma = 0)
}
\arguments{
\item{fit}{The output of L0Learn.fit}

\item{newx}{A matrix on which predictions are made. The matrix should have p columns.}

\item{lambda}{The value(s) of lambda to use for prediction. A summary of the lambdas in the regularization
path can be obtained using \code{print(fit)}.}

\item{gamma}{The value of gamma to use for prediction. A summary of the gammas in the regularization
path can be obtained using \code{print(fit)}.}
}
\description{
Predicts the response for a given sample
}
