% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.L0Learn}
\alias{print.L0Learn}
\title{Print L0Learn.fit object}
\usage{
\method{print}{L0Learn}(x, ...)
}
\arguments{
\item{x}{L0Learn.fit object}

\item{...}{ignore}
}
\description{
Prints a summary of L0Learn.fit
}
